+++ lm-facade-service/src/main/java/cz/fi/muni/legomanager/services/KitService.java
@@ -23,5 +23,5 @@ public interface KitService {
     public void decreaseBrickCountByOne(long kitId, long brickId);
     public void addBrickToKitById(long kitId, long brickId);
     public long getBrickCount(long kitId, long brickId);
-    public void setBrickCount(long kitId, long brickId, long amount);
+    public void setBrickCount(long kitId, long brickId, int amount);
 }
